+++ src/LaravelDpdServiceProvider.php
@@ -4,6 +4,7 @@ namespace AlexanderPoellmann\LaravelDpd;
 
 use AlexanderPoellmann\LaravelDpd\Contracts\DpdTransport;
 use AlexanderPoellmann\LaravelDpd\Services\RestTransport;
+use AlexanderPoellmann\LaravelDpd\Shipping\DpdShippingAdapter;
 use Spatie\LaravelPackageTools\Package;
 use Spatie\LaravelPackageTools\PackageServiceProvider;
 
@@ -20,5 +21,7 @@ class LaravelDpdServiceProvider extends PackageServiceProvider
     {
         $this->app->singleton(DpdTransport::class, RestTransport::class);
         $this->app->singleton(LaravelDpd::class);
+        $this->app->singleton(DpdShippingAdapter::class);
+        $this->app->tag([DpdShippingAdapter::class], 'shipping.adapters');
     }
 }